oonimkall: use external testing and better coverage
External testing is better because it explicits what is the API.

I would also like to have very good inter package coverage. In this
case, sadly, it means duplicating some tests.

I think there is more yak shaving to rationalize and speed up
tests within oonimkall a bit, but not immediately.

This finishes the yak shaving before fully tackling
https://github.com/ooni/probe-engine/issues/893.
