+++ oonimkall/tasks/runner_test.go
@@ -1,212 +1,32 @@
-package tasks
+package tasks_test
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"log"
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"sync"
 	"testing"
+	"time"
 
-	"github.com/google/go-cmp/cmp"
-	engine "github.com/ooni/probe-engine"
+	"github.com/ooni/probe-engine/oonimkall/tasks"
 )
-
-func TestUnitRunnerHasUnsupportedSettings(t *testing.T) {
-	out := make(chan *Event)
-	var falsebool bool
-	var zerodotzero float64
-	var zero int64
-	var emptystring string
-	settings := &Settings{
-		InputFilepaths: []string{"foo"},
-		Options: SettingsOptions{
-			AllEndpoints:          &falsebool,
-			Backend:               "foo",
-			BouncerBaseURL:        "https://ps-nonexistent.ooni.io/",
-			CABundlePath:          "foo",
-			CollectorBaseURL:      "https://ps-nonexistent.ooni.io/",
-			ConstantBitrate:       &falsebool,
-			DNSNameserver:         &emptystring,
-			DNSEngine:             &emptystring,
-			ExpectedBody:          &emptystring,
-			GeoIPASNPath:          "foo",
-			GeoIPCountryPath:      "foo",
-			Hostname:              &emptystring,
-			IgnoreBouncerError:    &falsebool,
-			IgnoreOpenReportError: &falsebool,
-			MLabNSAddressFamily:   &emptystring,
-			MLabNSBaseURL:         &emptystring,
-			MLabNSCountry:         &emptystring,
-			MLabNSMetro:           &emptystring,
-			MLabNSPolicy:          &emptystring,
-			MLabNSToolName:        &emptystring,
-			NoFileReport:          false,
-			Port:                  &zero,
-			ProbeASN:              "AS0",
-			ProbeCC:               "ZZ",
-			ProbeIP:               "127.0.0.1",
-			ProbeNetworkName:      "XXX",
-			RandomizeInput:        true,
-			SaveRealResolverIP:    &falsebool,
-			Server:                &emptystring,
-			TestSuite:             &zero,
-			Timeout:               &zerodotzero,
-			UUID:                  &emptystring,
-		},
-		OutputFilepath: "foo",
-	}
-	go func() {
-		defer close(out)
-		r := NewRunner(settings, out)
-		logger := NewChanLogger(r.emitter, "WARNING", r.out)
-		if r.hasUnsupportedSettings(logger) != true {
-			panic("expected to see unsupported settings")
-		}
-	}()
-	var fatal, warn []string
-	for ev := range out {
-		switch ev.Key {
-		case "failure.startup":
-			evv := ev.Value.(eventFailureGeneric)
-			if strings.HasSuffix("not supported", evv.Failure) {
-				log.Fatalf("invalid value: %s", evv.Failure)
-			}
-			fatal = append(fatal, evv.Failure)
-		case "log":
-			evv := ev.Value.(eventLog)
-			if strings.HasSuffix("not supported", evv.Message) {
-				log.Fatalf("invalid value: %s", evv.Message)
-			}
-			warn = append(warn, evv.Message)
-		default:
-			log.Fatalf("invalid key: %s", ev.Key)
-		}
-	}
-	expectedFatal := []string{
-		"InputFilepaths: not supported",
-		"Options.Backend: not supported",
-		"Options.BouncerBaseURL: not supported",
-		"Options.CollectorBaseURL: not supported",
-		"Options.Port: not supported",
-		"Options.RandomizeInput: not supported",
-		"Options.SaveRealResolverIP: not supported",
-		"Options.Server: not supported",
-		"Options.TestSuite: not supported",
-		"Options.Timeout: not supported",
-		"Options.UUID: not supported",
-		"OutputFilepath && !NoFileReport: not supported",
-	}
-	if diff := cmp.Diff(expectedFatal, fatal); diff != "" {
-		t.Fatal(diff)
-	}
-	expectedWarn := []string{
-		"Options.AllEndpoints: not supported",
-		"Options.CABundlePath: not supported",
-		"Options.ConstantBitrate: not supported",
-		"Options.DNSNameserver: not supported",
-		"Options.DNSEngine: not supported",
-		"Options.ExpectedBody: not supported",
-		"Options.GeoIPASNPath: not supported",
-		"Options.GeoIPCountryPath: not supported",
-		"Options.Hostname: not supported",
-		"Options.IgnoreBouncerError: not supported",
-		"Options.IgnoreOpenReportError: not supported",
-		"Options.MLabNSAddressFamily: not supported",
-		"Options.MLabNSBaseURL: not supported",
-		"Options.MLabNSCountry: not supported",
-		"Options.MLabNSMetro: not supported",
-		"Options.MLabNSPolicy: not supported",
-		"Options.MLabNSToolName: not supported",
-		"Options.ProbeASN: not supported",
-		"Options.ProbeCC: not supported",
-		"Options.ProbeIP: not supported",
-		"Options.ProbeNetworkName: not supported",
-	}
-	if diff := cmp.Diff(expectedWarn, warn); diff != "" {
-		t.Fatal(diff)
-	}
-}
-
-func TestUnitMeasurementSubmissionEventName(t *testing.T) {
-	if measurementSubmissionEventName(nil) != statusMeasurementSubmission {
-		t.Fatal("unexpected submission event name")
-	}
-	if measurementSubmissionEventName(errors.New("mocked error")) != failureMeasurementSubmission {
-		t.Fatal("unexpected submission event name")
-	}
-}
-
-func TestUnitMeasurementSubmissionFailure(t *testing.T) {
-	if measurementSubmissionFailure(nil) != "" {
-		t.Fatal("unexpected submission failure")
-	}
-	if measurementSubmissionFailure(errors.New("mocked error")) != "mocked error" {
-		t.Fatal("unexpected submission failure")
-	}
-}
-
-func TestIntegrationRunnerMaybeLookupLocationFailure(t *testing.T) {
-	out := make(chan *Event)
-	settings := &Settings{
-		AssetsDir: "../testdata/oonimkall/assets",
-		Name:      "Example",
-		Options: SettingsOptions{
-			SoftwareName:    "oonimkall-test",
-			SoftwareVersion: "0.1.0",
-		},
-		StateDir: "../testdata/oonimkall/state",
-	}
-	seench := make(chan int64)
-	go func() {
-		var seen int64
-		for ev := range out {
-			switch ev.Key {
-			case "failure.ip_lookup", "failure.asn_lookup",
-				"failure.cc_lookup", "failure.resolver_lookup":
-				seen++
-			case "status.progress":
-				evv := ev.Value.(eventStatusProgress)
-				if evv.Percentage >= 0.2 {
-					panic(fmt.Sprintf("too much progress: %+v", ev))
-				}
-			case "status.queued", "status.started", "status.end":
-			default:
-				panic(fmt.Sprintf("unexpected key: %s", ev.Key))
-			}
-		}
-		seench <- seen
-	}()
-	expected := errors.New("mocked error")
-	r := NewRunner(settings, out)
-	r.maybeLookupLocation = func(*engine.Session) error {
-		return expected
-	}
-	r.Run(context.Background())
-	close(out)
-	if n := <-seench; n != 4 {
-		t.Fatal("unexpected number of events")
-	}
-}
 
 func TestIntegrationRunnerMaybeLookupBackendsFailure(t *testing.T) {
 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(500)
 	}))
 	defer server.Close()
-	out := make(chan *Event)
-	settings := &Settings{
-		AssetsDir: "../testdata/oonimkall/assets",
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
 		Name:      "Example",
-		Options: SettingsOptions{
+		Options: tasks.SettingsOptions{
 			ProbeServicesBaseURL: server.URL,
 			SoftwareName:         "oonimkall-test",
 			SoftwareVersion:      "0.1.0",
 		},
-		StateDir: "../testdata/oonimkall/state",
+		StateDir: "../../testdata/oonimkall/state",
 	}
 	seench := make(chan int64)
 	go func() {
@@ -222,7 +42,7 @@ func TestIntegrationRunnerMaybeLookupBackendsFailure(t *testing.T) {
 		}
 		seench <- seen
 	}()
-	Run(context.Background(), settings, out)
+	tasks.Run(context.Background(), settings, out)
 	close(out)
 	if n := <-seench; n != 1 {
 		t.Fatal("unexpected number of events")
@@ -245,16 +65,16 @@ func TestIntegrationRunnerOpenReportFailure(t *testing.T) {
 		w.WriteHeader(500)
 	}))
 	defer server.Close()
-	out := make(chan *Event)
-	settings := &Settings{
-		AssetsDir: "../testdata/oonimkall/assets",
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
 		Name:      "Example",
-		Options: SettingsOptions{
+		Options: tasks.SettingsOptions{
 			ProbeServicesBaseURL: server.URL,
 			SoftwareName:         "oonimkall-test",
 			SoftwareVersion:      "0.1.0",
 		},
-		StateDir: "../testdata/oonimkall/state",
+		StateDir: "../../testdata/oonimkall/state",
 	}
 	seench := make(chan int64)
 	go func() {
@@ -264,7 +84,7 @@ func TestIntegrationRunnerOpenReportFailure(t *testing.T) {
 			case "failure.report_create":
 				seen++
 			case "status.progress":
-				evv := ev.Value.(eventStatusProgress)
+				evv := ev.Value.(tasks.EventStatusProgress)
 				if evv.Percentage >= 0.4 {
 					panic(fmt.Sprintf("too much progress: %+v", ev))
 				}
@@ -276,9 +96,322 @@ func TestIntegrationRunnerOpenReportFailure(t *testing.T) {
 		}
 		seench <- seen
 	}()
-	Run(context.Background(), settings, out)
+	tasks.Run(context.Background(), settings, out)
 	close(out)
 	if n := <-seench; n != 1 {
 		t.Fatal("unexpected number of events")
+	}
+}
+
+func TestIntegrationRunnerGood(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		LogLevel:  "DEBUG",
+		Name:      "Example",
+		Options: tasks.SettingsOptions{
+			SoftwareName:    "oonimkall-test",
+			SoftwareVersion: "0.1.0",
+		},
+		StateDir: "../../testdata/oonimkall/state",
+	}
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "status.end" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("status.end event not found")
+	}
+}
+
+func TestIntegrationRunnerWithUnsupportedSettings(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		LogLevel:  "DEBUG",
+		Name:      "Example",
+		Options: tasks.SettingsOptions{
+			SoftwareName:    "oonimkall-test",
+			SoftwareVersion: "0.1.0",
+		},
+		OutputFilepath: "/nonexistent",
+		StateDir:       "../../testdata/oonimkall/state",
+	}
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "failure.startup" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("failure.startup event not found")
+	}
+}
+
+func TestIntegrationRunnerWithInvalidKVStorePath(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		LogLevel:  "DEBUG",
+		Name:      "Example",
+		Options: tasks.SettingsOptions{
+			SoftwareName:    "oonimkall-test",
+			SoftwareVersion: "0.1.0",
+		},
+		StateDir: "/nonexistent/long/directory/name",
+	}
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "failure.startup" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("failure.startup event not found")
+	}
+}
+
+func TestIntegrationRunnerWithInvalidExperimentName(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		LogLevel:  "DEBUG",
+		Name:      "Nonexistent",
+		Options: tasks.SettingsOptions{
+			SoftwareName:    "oonimkall-test",
+			SoftwareVersion: "0.1.0",
+		},
+		StateDir: "../../testdata/oonimkall/state",
+	}
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "failure.startup" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("failure.startup event not found")
+	}
+}
+
+func TestIntegrationRunnerWithInconsistentGeolookupSettings(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		LogLevel:  "DEBUG",
+		Name:      "Example",
+		Options: tasks.SettingsOptions{
+			NoGeoIP:          true,
+			NoResolverLookup: false,
+			SoftwareName:     "oonimkall-test",
+			SoftwareVersion:  "0.1.0",
+		},
+		StateDir: "../../testdata/oonimkall/state",
+	}
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "failure.startup" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("failure.startup event not found")
+	}
+}
+
+func TestIntegrationRunnerWithNoGeolookup(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		LogLevel:  "DEBUG",
+		Name:      "Example",
+		Options: tasks.SettingsOptions{
+			NoGeoIP:          true,
+			NoResolverLookup: true,
+			SoftwareName:     "oonimkall-test",
+			SoftwareVersion:  "0.1.0",
+		},
+		StateDir: "../../testdata/oonimkall/state",
+	}
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "status.end" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("status.end event not found")
+	}
+}
+
+func TestIntegrationRunnerWithMissingInput(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		LogLevel:  "DEBUG",
+		Name:      "ExampleWithInput",
+		Options: tasks.SettingsOptions{
+			NoGeoIP:          true,
+			NoResolverLookup: true,
+			SoftwareName:     "oonimkall-test",
+			SoftwareVersion:  "0.1.0",
+		},
+		StateDir: "../../testdata/oonimkall/state",
+	}
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "failure.startup" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("failure.startup event not found")
+	}
+}
+
+func TestIntegrationRunnerWithMaxRuntime(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		Inputs:    []string{"a", "b", "c", "d", "e", "f", "g", "h", "i"},
+		LogLevel:  "DEBUG",
+		Name:      "ExampleWithInput",
+		Options: tasks.SettingsOptions{
+			MaxRuntime:       1,
+			NoGeoIP:          true,
+			NoResolverLookup: true,
+			SoftwareName:     "oonimkall-test",
+			SoftwareVersion:  "0.1.0",
+		},
+		StateDir: "../../testdata/oonimkall/state",
+	}
+	begin := time.Now()
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "status.end" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("status.end event not found")
+	}
+	// The runtime is long because of ancillary operations and is even more
+	// longer because of self shaping we may be performing (especially in
+	// CI builds) using `-tags shaping`). We have experimentally determined
+	// that ~10 seconds is the typical CI test run time. See:
+	//
+	// 1. https://github.com/ooni/probe-engine/pull/588/checks?check_run_id=667263788
+	//
+	// 2. https://github.com/ooni/probe-engine/pull/588/checks?check_run_id=667263855
+	if time.Now().Sub(begin) > 10*time.Second {
+		t.Fatal("expected shorter runtime")
+	}
+}
+
+func TestIntegrationRunnerWithMaxRuntimeNonInterruptible(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		Inputs:    []string{"a", "b", "c", "d", "e", "f", "g", "h", "i"},
+		LogLevel:  "DEBUG",
+		Name:      "ExampleWithInputNonInterruptible",
+		Options: tasks.SettingsOptions{
+			MaxRuntime:       1,
+			NoGeoIP:          true,
+			NoResolverLookup: true,
+			SoftwareName:     "oonimkall-test",
+			SoftwareVersion:  "0.1.0",
+		},
+		StateDir: "../../testdata/oonimkall/state",
+	}
+	begin := time.Now()
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "status.end" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("status.end event not found")
+	}
+	// The runtime is long because of ancillary operations and is even more
+	// longer because of self shaping we may be performing (especially in
+	// CI builds) using `-tags shaping`). We have experimentally determined
+	// that ~10 seconds is the typical CI test run time. See:
+	//
+	// 1. https://github.com/ooni/probe-engine/pull/588/checks?check_run_id=667263788
+	//
+	// 2. https://github.com/ooni/probe-engine/pull/588/checks?check_run_id=667263855
+	if time.Now().Sub(begin) > 10*time.Second {
+		t.Fatal("expected shorter runtime")
+	}
+}
+
+func TestIntegrationRunnerWithFailedMeasurement(t *testing.T) {
+	out := make(chan *tasks.Event)
+	settings := &tasks.Settings{
+		AssetsDir: "../../testdata/oonimkall/assets",
+		Inputs:    []string{"a", "b", "c", "d", "e", "f", "g", "h", "i"},
+		LogLevel:  "DEBUG",
+		Name:      "ExampleWithFailure",
+		Options: tasks.SettingsOptions{
+			MaxRuntime:       1,
+			NoGeoIP:          true,
+			NoResolverLookup: true,
+			SoftwareName:     "oonimkall-test",
+			SoftwareVersion:  "0.1.0",
+		},
+		StateDir: "../../testdata/oonimkall/state",
+	}
+	go func() {
+		tasks.Run(context.Background(), settings, out)
+		close(out)
+	}()
+	var found bool
+	for ev := range out {
+		if ev.Key == "failure.measurement" {
+			found = true
+		}
+	}
+	if !found {
+		t.Fatal("failure.measurement event not found")
 	}
 }
